ai-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.1
python-multipart>=0.0.18
google-genai>=1.0.0
openai-whisper>=20231117
pydantic>=2.0.0
pytest>=8.0.0
httpx>=0.27.0
